fix positive to be greater then 0
