+++ src/strategies/common.ts
@@ -1,4 +1,5 @@
 import { constrains, enumerations, specs } from "../data";
+import { Sign } from "../data/enumerations";
 import { random, randomIntInclusive } from "../random";
 import { SIGN_CHANGE } from "./constant";
 
@@ -11,6 +12,19 @@ export function getSign(sign?: enumerations.Sign): number {
   return random() < SIGN_CHANGE ? 1 : -1;
 }
 
+function minOrNegative(min: number, sign: Sign): number {
+  if (min > 0 && sign === Sign.Negative) {
+    return -1;
+  }
+  return min;
+}
+
+export function getMinBasedOnSign(min: number, sign: Sign): number {
+  if (min <= 0 && sign === Sign.Positive) {
+    return 1;
+  }
+  return minOrNegative(min, sign);
+}
 function getCorrectMaxMin(max: number, min: number): number[] {
   if (max < min) {
     return [min, max];
