+++ src/strategies/index.ts
@@ -1,14 +1,12 @@
 export {
   BooleanStrategy,
   DateStrategy,
-  FloatStrategy,
-  NumberStrategy,
   StrategyConstructor,
 } from "./base_strategies";
-
 export {
   EmailStrategy,
   StringStrategy,
   URLStrategy,
   UUIDStrategy,
 } from "./string_strategies";
+export { FloatStrategy, NumberStrategy } from "./number_strategies";
